+++ packages/policy-opa/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/policy
 
+## 1.0.103
+
+### Patch Changes
+
+- Updated dependencies [91c2128]
+- Updated dependencies [25a0447]
+- Updated dependencies [2cd80b3]
+- Updated dependencies [d06bb2a]
+- Updated dependencies [123d71f]
+- Updated dependencies [2fa5e0e]
+  - @ai-sdk/provider-utils@5.0.42
+  - ai@7.0.103
+  - @ai-sdk/provider@4.0.16
+
 ## 1.0.102
 
 ### Patch Changes